[Feat] : protocol Provider, class ProviderImpl 생성
ProviderImpl 구현

Related to: #6
