+++ GitHubSearchApp/Infrastructure/Endpoint/Endpoint.swift
@@ -24,3 +24,66 @@ struct Endpoint<R>: RequestResponsable {
         self.queryParameter = queryParameter
     }
 }
+
+enum APIEndpoints {
+    case searchRepo(text: String)
+}
+
+
+
+extension APIEndpoints {
+    var baseURL: String {
+        switch self {
+        default:
+            return "https://api.github.com/"
+        }
+    }
+    
+    var path: String {
+        switch self {
+        case .searchRepo:
+            return "/search/repositories"
+        }
+    }
+    
+    var queryParameter: Encodable {
+        switch self {
+        case .searchRepo(let text):
+            return SearchRepoRequestDTO(q: text)
+        }
+    }
+    
+    var responseType: Decodable.Type {
+        switch self {
+        case .searchRepo:
+            return SearchRepoResponseDTO.self
+        }
+    }
+}
+
+struct End<Info> {
+    
+}
+
+
+//extension APIEndpoints {
+//
+//    func getEndpoint() -> Endpoint<> {
+//        let type = responseType
+//        return Endpoint(baseURL: baseURL, path: path, queryParameter: queryParameter)
+//    }
+//}
+//
+//struct End {
+//
+//    let e: APIEndpoints
+//
+//    init(e: APIEndpoints) {
+//        self.e = e
+//    }
+//
+//    func getEndpoint() -> Endpoint<R> {
+//        return Endpoint<e.reponsetype>(baseURL: e.baseURL, path: e.path, queryParameter: e.queryParameter)
+//    }
+//
+//}
